# api/v1/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pakiparc Fleet Management API
  version: 1.0.0
  description: |
    RESTful API for Pakiparc fleet management system.

    ## Authentication
    All endpoints (except /auth/*) require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer YOUR_TOKEN_HERE
    ```

    ## Rate Limiting
    - 100 requests per minute per user/IP
    - Returns 429 status when limit exceeded

    ## Response Format
    All responses are JSON format with consistent structure:
    ```json
    {
      "success": true,
      "data": {...},
      "message": "Optional message"
    }
    ```

    Error responses:
    ```json
    {
      "error": true,
      "message": "Error description",
      "details": {...}
    }
    ```

  contact:
    name: Pakiparc Support
    email: support@pakiparc.com
    url: https://pakiparc.com

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.pakiparc.com/v1
    description: Production server
  - url: https://staging-api.pakiparc.com/v1
    description: Staging server
  - url: http://localhost/api/v1
    description: Development server
...
paths:
...
  /vehicles:
    get:
      tags:
        - Vehicles
      summary: List all vehicles
      description: Retrieve paginated list of vehicles for current company
      operationId: getVehicles
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, maintenance]
      responses:
        '200':
          description: List of vehicles
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vehicle'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Vehicles
      summary: Create vehicle
      description: Add a new vehicle to the fleet
      operationId: createVehicle
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleInput'
      responses:
        '201':
          description: Vehicle created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        registration_number:
          type: string
          example: ABC-123-XY
        make:
          type: string
          example: Renault
        model:
          type: string
          example: Master
        year:
          type: integer
          example: 2023
        fuel_type:
          type: string
          enum: [diesel, gasoline, electric, hybrid]
        status:
          type: string
          enum: [active, inactive, maintenance]
        current_mileage:
          type: integer
        purchase_date:
          type: string
          format: date
    VehicleInput:
      type: object
      required:
        - registration_number
        - make
        - model
        - fuel_type
      properties:
        registration_number:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        fuel_type:
          type: string
          enum: [diesel, gasoline, electric, hybrid]
        current_mileage:
          type: integer
        purchase_date:
          type: string
          format: date
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: boolean
                example: true
              message:
                type: string
                example: Unauthorized
...
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: boolean
                example: true
              message:
                type: string
              validation_errors:
                type: object